refactor(application): move bootstrap_factory from interfaces to application layer
Move bootstrap_factory module to application layer for better architectural alignment.
The interfaces layer now contains a deprecation stub that re-exports from application.
