+++ src/bioetl/application/bootstrap.py
@@ -16,8 +16,8 @@ Example:
     >>> context = bootstrap.start()
     >>> schema = context.schema_provider.get_schema("activity")
     >>>
-    >>> # With infrastructure integration (from interfaces layer)
-    >>> from bioetl.interfaces.bootstrap_factory import create_default_bootstrap
+    >>> # With infrastructure integration
+    >>> from bioetl.application.bootstrap_factory import create_default_bootstrap
     >>> bootstrap = create_default_bootstrap()
     >>> context = bootstrap.start()
 """
