+++ src/bioetl/interfaces/factories/infrastructure.py
@@ -47,7 +47,7 @@ class DefaultInfrastructureFactory(InfrastructureFactoryABC):
 
     def create_config_loader(self) -> PipelineConfigLoaderProtocol:
         """Create a pipeline configuration loader using bootstrap."""
-        from bioetl.interfaces.bootstrap_factory import create_default_bootstrap
+        from bioetl.application.bootstrap_factory import create_default_bootstrap
 
         bootstrap = create_default_bootstrap()
         context = bootstrap.start()
