+++ tests/integration/test_pipeline_e2e.py
@@ -13,7 +13,7 @@ from bioetl.application.mappers.chembl import ChemblRecordMapper
 from bioetl.infrastructure.clients.chembl.response_parser import (
     ChemblGenericResponseParser,
 )
-from bioetl.interfaces.bootstrap_factory import create_default_bootstrap
+from bioetl.application.bootstrap_factory import create_default_bootstrap
 
 
 class TestPipelineDataFlow:
